*: label the secrets we interact with
In order to reduce the cache footprint we have, we need to label the
secrets we interact with so we can use a label selector on the LIST +
WATCH requests we make. This commit does two things:

1. adds labels to all newly-created objects, so we have them moving
   forward
2. adds a controller that will react to the (unrestricted) set of
   secrets on the cluster today and puts labels where they are needed,
   so we can bring clusters without correct labels into conformance

A future release of the CCO will remove the second controller and
restrict the cache using our new label selector. We may need to do some
extra work like setting up new controllers to ensure users don't remove
labels from the admin secrets (or that we add labels to those when
they're new).

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>